refactor(api): URL-mirror layout, quality gates, and expanded tests
Migrate all Django apps off legacy api/ packages to path-aligned *_api modules,
generic class-based views, and config.api_urls aggregation. Split channels
helpers into services; relocate common tests; add tracks/playlists API tests,
OpenAPI snapshot export, pre-commit, Vitest hooks, and dashboard feature barrel.

+++ apps/api/apps/channels/services/party_recap.py
@@ -24,7 +24,9 @@ def _build_excitement_heatmap(channel_id: int, rows: list, *, buckets: int = 24)
             weight = 3.0 if ev in ("next", "skip", "vote_skip") else 2.0 if ev == "play" else 1.0
             points.append((row.emitted_at, weight))
 
-    chat_msgs = ChannelChatMessage.objects.filter(channel_id=channel_id, deleted_at__isnull=True).only("id", "created_at")[:500]
+    chat_msgs = ChannelChatMessage.objects.filter(channel_id=channel_id, deleted_at__isnull=True).only(
+        "id", "created_at"
+    )[:500]
     msg_ids = [m.id for m in chat_msgs]
     msg_times = {m.id: m.created_at for m in chat_msgs if m.created_at}
     if msg_ids:
@@ -54,7 +56,7 @@ def _build_excitement_heatmap(channel_id: int, rows: list, *, buckets: int = 24)
     max_score = max(bucket_scores) or 1.0
     out_buckets = []
     for i, sc in enumerate(bucket_scores):
-        pct = int(round(100 * sc / max_score)) if max_score else 0
+        pct = round(100 * sc / max_score) if max_score else 0
         label_min = int((i / buckets) * span // 60)
         out_buckets.append(
             {
